finish post unit test

+++ src/posts/services/post.service.ts
@@ -20,12 +20,13 @@ export class PostService {
 
   async getPosts(): Promise<TPost[]> {
     const posts = await this.postRepo.find({ order: { createdAt: 'ASC' } })
-    if (posts.length !== 0) throw new NotFoundException('Posts not found')
+    if (posts.length === 0) throw new NotFoundException('Posts not found')
     return plainToClass(TPost, posts)
   }
 
   async getPost(id: number): Promise<TPost> {
     const post = await this.postRepo.findOne(id)
+    if (!post) throw new NotFoundException('Post not found')
     return plainToClass(TPost, post)
   }
 
@@ -34,15 +35,17 @@ export class PostService {
       where: { topicId },
       order: { createdAt: 'ASC' },
     })
+    if(posts.length === 0) throw new NotFoundException('Posts not found')
     return plainToClass(TPost, posts)
   }
 
   async getFirstPostByTopicId(topicId: number): Promise<TPost> {
-    const posts = await this.postRepo.findOne({
+    const post = await this.postRepo.findOne({
       where: { topicId },
       order: { createdAt: 'ASC' },
     })
-    return plainToClass(TPost, posts)
+    if(!post)throw new NotFoundException('Post not found')
+    return plainToClass(TPost, post)
   }
 
   async createPost(data: IPost): Promise<PostEntity> {
@@ -55,11 +58,14 @@ export class PostService {
   }
 
   async deletePosts(topicId: number): Promise<DeleteResult> {
+    const postsData = await this.getPostsByTopicId(topicId)
+    if(postsData.length === 0)throw new NotFoundException('Posts not found')
     const deletedPosts = this.postRepo.delete({ topicId })
     return deletedPosts
   }
 
   async deletePost(postId: number): Promise<DeleteResult> {
+    
     const deletedPost = this.postRepo.delete({ id: postId })
     return deletedPost
   }
